rename: rename configuarions in backend server
