+++ backend/app/core/settings.py
@@ -18,4 +18,3 @@ class Settings(BaseSettings):
 
 
 settings = Settings()
-
